requirements.txt
Flask==3.0.2
PyPDF2==3.0.1
python-docx==1.1.0
spacy==3.8.11
scikit-learn==1.8.0
pandas==2.3.2

# After installing, run: python -m spacy download en_core_web_sm

